refactor: rename options modules to settings
