+++ test/optionsHelpers.test.ts
@@ -2,7 +2,7 @@
 
 (window as any).electron = { getBaseDir: () => Promise.resolve(__dirname) };
 
-import { _test } from '../app/ts/renderer/options';
+import { _test } from '../app/ts/renderer/settings';
 import { settings } from '../app/ts/renderer/settings-renderer';
 import appDefaults from '../app/ts/appsettings';
